enable service name prefix for datadog
To allow the datadog stats recorder to work with the set service name/namespace, we need to add the prefix to the key. We do this by mimicing what armon/go-metrics metrics class does, but separated from the calls to the individual sink to send to datadog with a tag.